license: add GPLv2 COPYING, DISCLAIMER, headers
Verified paper records that M/Linux 1.0 was GPLv2 under
UCRL-CODE-2000-010, (c) 2000 Regents of the University of California.
Modifications to DISCLAIMER for GPL were approved in writing.

The release was approved on 1999-12-27 and GPL release instructions
were clarified on 2000-02-23.

That being the case, for some reason the appropriate DISCLAIMER,
COPYING, and source headers were not present in the source I found
in my archive.  Although that might indicate it is not the latest
version, I went ahead and added the headers to this version.